Face Reconstruction/Landmark Detection and 3D Face Reconstruction for Caricature using a Nonlinear Parametric Model/requirements.txt
numpy==1.18.1
torch==1.4.0
torchvision==0.5.0
pillow==7.0.0
opencv-python==3.4.2.17